harpoon: add `undo` command
This patch adds a history system to Harpoon backed by an OCaml
`Stack.t`. The history items are instances of a new concept called a
`Theorem.Action.t`. This is essentially a diff on the subgoal list
that results from the execution of a proof tactic. All changes to a
theorem's subgoal list have to go through the `Theorem.apply` function
that takes an `Action.t` and records it internally in the theorem.
Actions can be run backwards, which is exactly how undo is
implemented.