doc: undo; stub inductive types & proof scripts
Also, I restructured the contextual LF page a bit. It might be worth
splitting this page up a bit since it spends a lot of time
discussing Beluga computation types.  This discussion might be
better moved to a page called "Beluga type system" so that the
discussion of contextual LF can concern itself primarily with
metavariables.