synint: track plicity of MetaApp
Similar to MApp in terms, MetaApp in inductive type constructor spines
may be implicitly reconstructed, when the reconstruction target type
is an implicit PiBox kind. The MetaApp constructor now remembers
whether it was implicitly reconstructed or explicitly supplied by the
programmer, so that during pretty-printing, we can avoid printing the
implicitly reconstructed applications.