harpoon: fix auto_intros generating duplicate variable names
Instead of using the names in the given mctx and gctx as the names to
avoid in `intros'`, I added an argument to specify the initial list of
names to avoid. This way, auto-intros (which passes in empty contexts)
can specify the names present in the actual, possibly nonempty
contexts. This avoids the previously present bug that we would end up
with `x` several times in the program context.