harpoon: add multiple subgoals in two phases
First add all the subgoals. Then run all the subgoal hooks.
This way, the only time Harpoon will prevent an auto_solve_trivial
from running after splitting is if auto_solve_trivial is capable of
solving *all* of the generated subgoals.

Also when splitting, the first generated subgoal is often the case for
a parameter variable for which a projection is capable of solving the
subgoal. So we get a better user experience.

Fixes #206