reconstruct: add extra calls to normalization
This is necessary to replace instantiated MMVars with their
instantiations so that typechecking Harpoon proofs works later.

In Beluga, this is dealt with by just normalizing the entire
reconstructed program, but I never bothered implemented a `cnorm*`
family of functions for Harpoon proofs since they don't admit general
contextual normalization.