unify: more careful plicity management
When unifying two MMVars with itself, we construct a new MMVar that
depends on the intersection of the dependencies of the two input
MMVars. Then we instantiate both input MMVars to the newly built
MMVar.

The problem is to decide what plicity to assign to the new MMVar!
Beluga was being very conservative in always constructing an implicit
MMVar. But if we instantiate an explicit MMVar with an implicit MMVar,
this could have the effect of making an explicitly-bound variable go
out of scope!

This patch introduces a somewhat hamfisted way of resolving this
issue. If both MMVars are explicit, then the new MMVar they get
instantiated with will also be explicit. This at least resolves some
known issues where variables go out of scope for no apparent
reason. In reality it's because the variable ends up being unified
with itself.

Closes #213