reconstruct harpoon: improve contextual name generation
When reconstructing a comp split, I had assumed that the only possible
variables being introduced would be program variables. But this
assumption is wrong: when a data constructor has explicit PiBox types,
these MUST appear as meta-objects in the pattern. Of course, these
meta-objects will contain metavariables. So when performing contextual
name generation, we also need to consider the names in cD.