unify constraints: add conv check for IHead and ISub
Constraint solving checked convertibility of terms for INorm
equations, but not for IHead and ISub equations. This patch extends
constraint solving to include convertibility checking for those kinds
of equations as well, since we can end up with trivial equations
between MPVars (which are IHead) that otherwise cannot be solved
because they fall outside the pattern fragment.