unify: drop solved constraints as we go
It could happen that instantiating an MMVar would add its local
constraints to the global list and that those constraints involve the
MMVar itself. Since this is happening during constraint solving, the
local constraints are actually queued already, i.e. they've been
solved. But then these solved constraints do not get cleared, so the
solver believes that the solution introduced new constraints / readded
the same constraint again, which is an error.

Now the procedure for solving one constraint is:
- unify the sides of the constraint
- try to solve any constraints that it could generate
- check that all global constraints are solved
  -> if not, that's an error
- ***clear the global constraint list***
  this is the part I added