TEST: just check status code
Harpoon will exit with `1` when running in `--stop` mode if an error
occurs at any point.

The previous technique checks the actual text output of Harpoon for
certain bad strings and returns 1 in that case, which does not
cooperate well with setups involving multiple theorems or in certain
situations where the bad strings don't appear despite an error
occurring.

Fixes #216