fix recover_theorem; refactor Total.annotate
Both theorem.ml and recsgn.ml needed to apply totality
annotations. This actually required quite the song-and-dance in both
places, so now that logic is centralized in total.ml, replacing the
old `annotate` function that was lower-level.

Theorem recovery is now made to construct the initial subgoal with the
annotated theorem statement. This is what the translation expects.

Fixes #215, #214.