total: let comp holes participate in useIH
Computational holes are now allowed to fill in any kind of requirement
for a recursive call.
An LFHole appearing in a box, however, will cause an error about the
recursive call being not structurally smaller. In that case, the
workaround is to simply comment out the totality declaration, complete
the proof, and then add it back.

Closes #77