automation: fix dctx_hat generated in auto-solve-trivial
The erased LF context constructed for a metavariable trivial witness
was not being appropriately shifted. When came the time to later
translate and check this term, its context variable was referring to
something that wasn't a context variable in Delta!

I'm very surprised that this bug took this long to notice, since it
seems like something that would completely break metavariable trivial
witnesses.

Closes #229