Ensure test files can be written to
Newer versions of `dune` make it so that the built files are read-only.
This would cause some tests to fail because these files are be copied in a temporary directory with the same permissions, and they were not assumed to be read-only.
A `chmod` is done recursively on the temporary directory to fix this.