Use `Format.asprintf` where `Format.fprintf str_formatter` and `flush_str_formatter` were used
Use of the global `Format.stdbuf` buffer is not safe unless it is flushed after each use.
The changed areas mainly consist in error message pretty-printing to string.

The existing architecture where error messages are generated when they are raised does not allow for parameterising the formatter used for printing.
Addressing this issue is outside the scope of this change.