Re-implement parsing of weak functional space LF kinds, types and terms
The LF and contextual LF parsers are split into distinct modules.
This is because LF kinds, types and terms used in LF definitions cannot contain substitutions.

This parser also supports backward arrows.
It remains to unit-test these parsers, and this will require a revised external syntax as the target of data-dependent disambiguation.