Fix bug where prefix arguments would be elaborate to an application
The elimination of juxtaposed objects was too eager.
Type applications such as `sum x y z` with type-level constant `sum` and term-level variables `x`, `y`, `z` was being elaborated as `App(sum, []) App(x, [y, z])`, leading to `ShuntingYard.Leftover_expressions` being raised.
The solution is to greedily take all arguments if the head operator is prefix.
This has the consequence that `s z` with type-level constant `s` and term-level constant `z` is parsed as `App(s, [z])` instead of `App(s, [App(z, [])])`.
This consequence is more consistent with quoted operators which are not elaborated to applications.