Fix handling of pattern variables
Handling of plain identifiers was the same as for elaborating terms and types, whereby exceptions would be raised if the variable was bound and of the wrong sort.

Now, in this elaboration, pattern variables are not required to be unique.
However, elaborating pattern-matching branches will require that pattern variables be unique, otherwise we may not correctly elaborate term applications with variable heads.