Remove explicit parentheses from the model for the external syntax
In order to simplify pretty-printing of the external syntax, the `*Parenthesized` constructors have been removed.
This has the consequence that the rewriting of backward arrows to be left-associative fails in cases where parentheses were required.
The expression parser from `parsec` is used instead to parse static left-associative operators in `Parser`.