Fix locations in applications
The `Location.join` function did not behave as a `union` of locations.
That is, `Location.join l1 l2` is not necessarily equal to `Location.join l2 l1`.
This used to be only in use for the parser, but disambiguation requires taking the union of locations that are out of order.

The `Location.between` function was introduced specifically for the parser.