Allow free context variables
In the previous parser, an untyped binding in head position of a context is a free or bound context variable.
That is, `g` is a free context variable in `[g]` if `g` is unbound.
If `g` is a bound context variable, then `[g]` is disambiguated the same way.

This makes it impossible to have multiple context variables in a context.