Fix parsing of declarations terminated by `.`
Old style LF declarations and pragmas are terminated by `.`.
Old style LF declarations may appear consecutively, as in:
```
tp : type.
base : tp.
arr : tp → tp → tp.
```
With the lexical convention that `x.y` is tokenized as `IDENT "x"; DOT; IDENT "y"`, we  end up with `tp.arr` being parsed as a qualified identifier.

The `DOT_IDENT` and `DOT_INTLIT` tokens are reintroduced to fix this.
This adds a whitespace sensitivity specifically for those declarations.
A hack is introduced in the parser to insert tokens in the input stream.
This allows for `DOT_IDENT` and `DOT_INTLIT` to be re-tokenized as `DOT; IDENT` and `DOT; INTLIT` when parsing a dot.