Use the state monad for HTML printing
This makes it easier to manage the state of qualified identifiers for pretty-printing.
It is arguably more readable to have the `(++)` append combinator than to have a format string followed by the list of printers and arguments to substitute in the format string.
This implementation also ensures that all boxes are correctly closed, and that parentheses are matched.