Implement stateful pretty-printing and fix the disambiguation of applications
During code review, it was noted that the concrete syntax AST should not contain the fields `operator` and `prefixed` which were populated during the disambiguation phase.
Instead, this data should be reconstructed by the pretty-printer.

During this refactoring/rewrite, it also became apparent that parsing/disambiguation for applications was not defined properly.
Indeed, there is no such thing as a non-unary prefix operator.
The application disambiguation implementation using the shunting yard algorithm could not handle `¬ ¬ ¬ A` with unary prefix operator `¬` because it would be parsed as the application of `¬` with arguments `{¬, ¬, A}`.
That is, the user was required to write `¬(¬(¬A))`.

A dynamic recursive descent parser based on Parsec's expression parser builder was implemented to remedy this.

Breaking changes:
- The Beluga `--prefix` pragma may only be used on unary constants