Simplify `Store.FVar`
It is never the case that `Store.FVar.add x tA` is called with `tA = TypVar _`.
It is hence possible to simplify `Store.FVar` to be a simple hash map.

It is also never the case that `Store.FVar.add x tA` is called with `x` already present in the store.
Had that been the case, `Store.FVar.add x tA` would fail with an inexhaustive pattern-matching exception.