Fix bugs in parsing and pretty-printing
A space and a hash were missing in printing of totality declarations and Harpoon split branch labels respectively.

Additionally, the case label `#1.2:` could not be parsed correctly because the lexical convention allowed identifiers to start with `#` (and `$`).
This means that the label was parsed as the identifier `"#1"` with an unexpected projection `.2`.