Fix test runner
The `run_harpoon_test.sh` script was only used in `TEST.sh`, so it was inlined.

Using `make test` did not always trigger a new build of the executables.

If `$PATH` had a `beluga` program, it would use that one instead of the newly built one.
This is fixed by re-introducing the environment variables for the executables, and using `dune exec which ...` to find their paths.

The interactive mode tests suppressed errors because `local output=$(...)` always has an exit code of `0`, whereas
```
local output
output=$(...)
```
does not.