Remove legacy handling of non-sequential scope-checking
The legacy system did not fully work because indexing/disambiguation was not implemented statefully with respect to a `Decl.t`.
As such, only the latest declarations could be used for all Harpoon subgoal holes, irrespective of their position in the signature.

The new system requires that the referencing environment be known before disambiguation and indexing, which ensures that valid commands respect the scope.