Re-add the `Depend` flag
There is a lot of confusion with the various flags used to annotate parameters in dependent function types and context bindings.

In particular, this `depend` flag was conflated with `plicity` and `inductivity` at some point during development.

- `depend` is strictly for dependency information between `x` and `B` in `Πx:A.B`
- `plicity` is strictly for explicit Pi-kinds like `{x : U} K` and implicit Pi-kinds like `(x : U) K`
- `inductivity` is strictly for totality-checking, to generate induction hypotheses

An addition `reconstructed` flag may later be required to correctly implement erasure during pretty-printing of the internal syntax.