Fix failing Harpoon tests
A hack was introduced to replace fresh identifiers generated by the indexing phase `"i` to `x` during pretty-printing of the internal syntax.

The pretty-printer for the internal syntax did not correctly add the now mandatory `#` and `$` prefixes to parameter types/objects and substitution types/objects respectively.

Parsing of weak prefix computation-level types as the last argument of an application would break parsing of Harpoon `suffices` directives followed by `intros`.

Some induction hypotheses are broken, likely because of the inconsistent uses of `Plicity.Explicit` and `Plicity.Implicit`.