Add test case for postponed infix pragmas
This test case is derived from `./t/code/success/infix/alternate.bel`.
The difference is that the new style of LF declarations is used, along with infix pragmas that precedence the type family declaration.
Those pragmas are postponed, in the sense that they apply to constants declared after it as opposed to constants already in scope.