Only add the latest bindings in `Binding_tree.add_all`
With `Hashtabl.{iter,fold}`, bindings are iterated over in reverse order of addition.
This means that  `Identifier.Hashtbl.iter (fun identifier entry -> Identifier.Hashtbl.add t1 identifier entry) t2` is such that the oldest binding for an identifier is added last.

This explains a bug with opening modules, whereby if a module defines overloaded constants, then the first definition is brought forward in scope as opposed to the last definition.
The fix is simlar to what is done for binding tree snapshots.