Rename "scope" operations to "frame" operations
A frame is a lexical region of an AST wherein variable declarations and references can be made.
A scope is a frame in which name resolution is uniform, meaning that references are resolved to the same declaration site irrespective of the AST node in the scope.

The current implementation for name resolution uses frames, not scopes.