Add support for case studies as tests
Case studies are tested just like examples, but their results are output in a separate group.
Case studies are also added as part of the pretty-printer test cases.
The file `./case-studies/PLACEHOLDER` can be deleted once an actual case study is added to the repository.

Support file paths with spaces for all test cases

In Bash, `for-in` loops split on `\t`, `\n` and ` `, so file paths containing spaces would be split.
Using `read -r file_path` fixes this.