Use `timeout` command in `TEST.sh`
The Github Actions CI pipeline fails on MacOS because `ulimit -t "10s"` unexpectedly causes the error `Cputime limit exceeded: 24`.
This is likely because of breaking changes with the upgraded MacOS image `12.7.3 -> 14.4.1`.

The `timeout` command is used instead to ensure the timeout applies to each test case separately.
This does increase the reported test runtime since extra sub-processes have to be spawned specifically for that `timeout` command to work.

`coreutils` is now required on MacOS to run `TEST.sh`.
It can be installed using `brew install coreutils`.