Promote `.out` files when `RESET_OUT_FILES=true`
In order for the integration tests to not affect the files in the repository,  test case files are first copied to a `TEMPDIR`, and then Beluga or Harpoon is executed on them.

When `RESET_OUT_FILES=true`, we expect the `.out` test files to be regenerated.
These used to be written in the `TEMPDIR`.
Now, they are written in `ROOTDIR`, which is the path to the repository.